+++ src/accessories/AccessoryService.ts
@@ -19,7 +19,10 @@ import { UnitLightbulb } from './UnitLightbulb';
 import { Thermostat } from './Thermostat';
 import { EmergencyAlarmSwitch } from './EmergencyAlarmSwitch';
 import { LockMechanism } from './LockMechanism';
+import { TemperatureSensor} from './TemperatureSensor';
+import { HumiditySensor } from './HumiditySensor';
 import { EmergencyTypes } from '../omni/messages/enums';
+import { SensorTypes } from '../omni/messages/AuxiliarySensorPropertiesResponse';
 
 export class AccessoryService {
   private accessories: Map<string, AccessoryBase> = new Map();
@@ -48,6 +51,8 @@ export class AccessoryService {
       this.discoverThermostats();
       this.discoverEmergencyAlarmSwitches();
       this.discoverAccessControls();
+      this.discoverTemperatureSensors();
+      this.discoverHumiditySensors();
     } catch (error) {
       this.platform.log.error(error);
     }
@@ -521,6 +526,58 @@ export class AccessoryService {
     }
   }
 
+  discoverTemperatureSensors(): void {
+    this.platform.log.debug(this.constructor.name, 'discoverTemperatureSensors');
+
+    const temperatureSensors = new Map<number, string>();
+
+    if (this.platform.settings.includeAuxiliarySensors) {
+      for(const [index, auxiliarySensor] of this.platform.omniService.auxiliarySensors) {
+        if (auxiliarySensor.sensorType === SensorTypes.Temperature) {
+          temperatureSensors.set(index, auxiliarySensor.name);
+        }
+      }
+    }
+
+    for(const [index, name] of temperatureSensors) {
+      this.addAccessory(TemperatureSensor, TemperatureSensor.type, name, index);
+    }
+
+    for(const accessory of this.accessories.values()) {
+      if (accessory instanceof TemperatureSensor) {
+        if (!temperatureSensors.has(accessory.platformAccessory.context.index)) {
+          this.removeAccessory(TemperatureSensor.type, accessory.platformAccessory.context.index);
+        }
+      }
+    }
+  }
+
+  discoverHumiditySensors(): void {
+    this.platform.log.debug(this.constructor.name, 'discoverHumiditySensors');
+
+    const humiditySensors = new Map<number, string>();
+
+    if (this.platform.settings.includeAuxiliarySensors) {
+      for(const [index, auxiliarySensor] of this.platform.omniService.auxiliarySensors) {
+        if (auxiliarySensor.sensorType === SensorTypes.Humidity) {
+          humiditySensors.set(index, auxiliarySensor.name);
+        }
+      }
+    }
+
+    for(const [index, name] of humiditySensors) {
+      this.addAccessory(HumiditySensor, HumiditySensor.type, name, index);
+    }
+
+    for(const accessory of this.accessories.values()) {
+      if (accessory instanceof HumiditySensor) {
+        if (!humiditySensors.has(accessory.platformAccessory.context.index)) {
+          this.removeAccessory(HumiditySensor.type, accessory.platformAccessory.context.index);
+        }
+      }
+    }
+  }
+
   addAccessory<TAccessory extends AccessoryBase>(
     Accessory: new (platform: OmniLinkPlatform, accessory: PlatformAccessory) => TAccessory,
     type: string, name: string, index?: number,
@@ -614,6 +671,10 @@ export class AccessoryService {
         return new EmergencyAlarmSwitch(this.platform, platformAccessory);
       case 'lockmechanism':
         return new LockMechanism(this.platform, platformAccessory);
+      case 'temperaturesensor':
+        return new TemperatureSensor(this.platform, platformAccessory);
+      case 'humiditysensor':
+        return new HumiditySensor(this.platform, platformAccessory);
     }
   }